Fixed most issues with login and creating user
Within Database.php there were a bunch of tabs at the start of the file. This was actually throwing off your ajax call. It was processing the tab spaces and the text it was outputting which was causing issues. Some mini map issues were fixed but there might be a few still remaining that you will have to look into.